fixed x64-windows-static triplet

+++ ports/ace/portfile.cmake
@@ -303,7 +303,7 @@ if(VCPKG_TARGET_IS_WINDOWS)
 
   # Remove dlls without any export
   if("tao" IN_LIST FEATURES OR "xml" IN_LIST FEATURES)
-    if (VCPKG_LIBRARY_LINKAGE STREQUAL dynamic)
+    if(VCPKG_LIBRARY_LINKAGE STREQUAL dynamic)
       file(REMOVE
         ${CURRENT_PACKAGES_DIR}/bin/ACEXML_XML_Svc_Conf_Parser.dll
         ${CURRENT_PACKAGES_DIR}/bin/ACEXML_XML_Svc_Conf_Parser.pdb
@@ -314,16 +314,18 @@ if(VCPKG_TARGET_IS_WINDOWS)
 
   # remove (erroneous) duplicate libs *TODO*: where do these come from ?
   if("tao" IN_LIST FEATURES)
-    file(REMOVE
-      ${CURRENT_PACKAGES_DIR}/debug/lib/tao_cosconcurrency.lib
-      ${CURRENT_PACKAGES_DIR}/debug/lib/tao_cosevent.lib
-      ${CURRENT_PACKAGES_DIR}/debug/lib/tao_coslifecycle.lib
-      ${CURRENT_PACKAGES_DIR}/debug/lib/tao_cosnaming.lib
-      ${CURRENT_PACKAGES_DIR}/debug/lib/tao_cosnotification.lib
-      ${CURRENT_PACKAGES_DIR}/debug/lib/tao_costrading.lib
-      ${CURRENT_PACKAGES_DIR}/debug/lib/tao_imr_activator.lib
-      ${CURRENT_PACKAGES_DIR}/debug/lib/tao_imr_locator.lib
-      ${CURRENT_PACKAGES_DIR}/debug/lib/tao_rtevent.lib)
+    if(VCPKG_LIBRARY_LINKAGE STREQUAL dynamic)
+      file(REMOVE
+        ${CURRENT_PACKAGES_DIR}/debug/lib/tao_cosconcurrency.lib
+        ${CURRENT_PACKAGES_DIR}/debug/lib/tao_cosevent.lib
+        ${CURRENT_PACKAGES_DIR}/debug/lib/tao_coslifecycle.lib
+        ${CURRENT_PACKAGES_DIR}/debug/lib/tao_cosnaming.lib
+        ${CURRENT_PACKAGES_DIR}/debug/lib/tao_cosnotification.lib
+        ${CURRENT_PACKAGES_DIR}/debug/lib/tao_costrading.lib
+        ${CURRENT_PACKAGES_DIR}/debug/lib/tao_imr_activator.lib
+        ${CURRENT_PACKAGES_DIR}/debug/lib/tao_imr_locator.lib
+        ${CURRENT_PACKAGES_DIR}/debug/lib/tao_rtevent.lib)
+    endif()
   endif()
 
   vcpkg_clean_msbuild()
